third_party/WebKit/LayoutTests/transitions/unprefixed-transform-expected.txt

-webkit-transform 10ms
fired transitionend with propertyName: -webkit-transform

transform 10ms
fired transitionend with propertyName: transform

transform 10ms, -webkit-transform 10ms
fired transitionend with propertyName: -webkit-transform

-webkit-transform 10ms, transform 10ms
fired transitionend with propertyName: transform

all 10ms
fired transitionend with propertyName: transform

all 10ms, -webkit-transform 10ms
fired transitionend with propertyName: -webkit-transform

-webkit-transform 10ms, all 10ms
fired transitionend with propertyName: transform
